renamed CellState to cell, moved piece and cell to data package
