+++ frontend/composeApp/src/commonMain/kotlin/pw/checkers/data/Cell.kt
@@ -1,0 +1,10 @@
+package pw.checkers.data
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Cell(
+    val row: Int,
+    val col: Int,
+    val piece: Piece? = null,
+)
