+++ frontend/composeApp/src/commonMain/kotlin/pw/checkers/data/Piece.kt
@@ -1,0 +1,12 @@
+package pw.checkers.data
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+enum class PieceType() {
+    PAWN, QUEEN
+}
+
+@Serializable
+data class Piece(val color: PlayerColor, val type: PieceType)
+
