+++ frontend/composeApp/src/commonMain/kotlin/pw/checkers/models/InitialBoard.kt
@@ -1,15 +1,18 @@
 package pw.checkers.models
 
-import pw.checkers.util.PlayerColor
+import pw.checkers.data.Cell
+import pw.checkers.data.Piece
+import pw.checkers.data.PieceType
+import pw.checkers.data.PlayerColor
 
-typealias Board = List<List<CellState>>
+typealias Board = List<List<Cell>>
 
 fun createInitialBoard(): Board = List(8) { row ->
     List(8) { col ->
         when {
-            row < 3 && (row + col) % 2 == 1 -> CellState(row, col, Piece(PlayerColor.BLACK, PieceType.PAWN))
-            row > 4 && (row + col) % 2 == 1 -> CellState(row, col, Piece(PlayerColor.WHITE, PieceType.PAWN))
-            else -> CellState(row, col)
+            row < 3 && (row + col) % 2 == 1 -> Cell(row, col, Piece(PlayerColor.BLACK, PieceType.PAWN))
+            row > 4 && (row + col) % 2 == 1 -> Cell(row, col, Piece(PlayerColor.WHITE, PieceType.PAWN))
+            else -> Cell(row, col)
         }
     }
 }
